+++ bundles/AdminBundle/Security/User/User.php
@@ -15,6 +15,16 @@
 
 namespace Pimcore\Bundle\AdminBundle\Security\User;
 
+use Pimcore\Security\User\User as PimcoreUser;
+
+trigger_deprecation(
+    'pimcore/pimcore',
+    '10.6',
+    'The "%s" class is deprecated and will be removed in Pimcore 11. Use %s instead.',
+    [User::class, PimcoreUser::class]
+);
+
+
 /**
  * @deprecated and will be removed in Pimcore 11. Use \Pimcore\Security\User\User instead.
  */
