+++ lib/Controller/UserAwareController.php
@@ -13,20 +13,15 @@
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\CoreBundle\Controller;
+namespace Pimcore\Controller;
 
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Pimcore\Security\User\User as UserProxy;
-use Pimcore\Controller\Controller;
 use Pimcore\Extension\Bundle\PimcoreBundleManager;
 use Pimcore\Logger;
 use Pimcore\Model\User;
 use Pimcore\Security\User\TokenStorageUserResolver;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
-use Symfony\Component\Serializer\Encoder\DecoderInterface;
-use Symfony\Component\Serializer\SerializerInterface;
-use Symfony\Component\Translation\Exception\InvalidArgumentException;
 use Symfony\Contracts\Service\Attribute\Required;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
@@ -36,28 +31,6 @@ abstract class UserAwareController extends Controller
      * @var TokenStorageUserResolver|TokenStorageUserResolver
      */
     protected $tokenResolver;
-
-    /**
-     * @var TranslatorInterface
-     */
-    protected $translator;
-
-    /**
-     * @var PimcoreBundleManager
-     */
-    protected $bundleManager;
-
-    #[Required]
-    public function setTranslator(TranslatorInterface $translator): void
-    {
-        $this->translator = $translator;
-    }
-
-    #[Required]
-    public function setBundleManager(PimcoreBundleManager $bundleManager): void
-    {
-        $this->bundleManager = $bundleManager;
-    }
 
     #[Required]
     public function setTokenResolver(TokenStorageUserResolver $tokenResolver): void
@@ -178,95 +151,5 @@ abstract class UserAwareController extends Controller
         if (null === $actionName || !in_array($actionName, $unrestrictedActions)) {
             $this->checkPermission($permission);
         }
-    }
-
-    /**
-     * Returns a JsonResponse that uses the admin serializer
-     *
-     * @param mixed $data    The response data
-     * @param int $status    The status code to use for the Response
-     * @param array $headers Array of extra headers to add
-     * @param array $context Context to pass to serializer when using serializer component
-     * @param bool $usePimcoreSerializer
-     *
-     * @return JsonResponse
-     *
-     */
-    protected function jsonResponse($data, $status = 200, $headers = [], $context = [], bool $usePimcoreSerializer = true)
-    {
-        $json = $this->encodeJson($data, $context, JsonResponse::DEFAULT_ENCODING_OPTIONS, $usePimcoreSerializer);
-
-        return new JsonResponse($json, $status, $headers, true);
-    }
-
-    /**
-     * Encodes data into JSON string
-     *
-     * @param mixed $data    The data to be encoded
-     * @param array $context Context to pass to serializer when using serializer component
-     * @param int $options   Options passed to json_encode
-     * @param bool $usePimcoreSerializer
-     *
-     * @return string
-     */
-    protected function encodeJson($data, array $context = [], $options = JsonResponse::DEFAULT_ENCODING_OPTIONS, bool $usePimcoreSerializer = true)
-    {
-        /** @var SerializerInterface $serializer */
-        $serializer = null;
-
-        if ($usePimcoreSerializer) {
-            $serializer = $this->container->get('pimcore.serializer');
-        } else {
-            $serializer = $this->container->get('serializer');
-        }
-
-        return $serializer->serialize($data, 'json', array_merge([
-            'json_encode_options' => $options,
-        ], $context));
-    }
-
-    /**
-     * Decodes a JSON string into an array/object
-     *
-     * @param mixed $json       The data to be decoded
-     * @param bool $associative Whether to decode into associative array or object
-     * @param array $context    Context to pass to serializer when using serializer component
-     * @param bool $usePimcoreSerializer
-     *
-     * @return mixed
-     */
-    protected function decodeJson($json, $associative = true, array $context = [], bool $usePimcoreSerializer = true)
-    {
-        /** @var SerializerInterface|DecoderInterface $serializer */
-        $serializer = null;
-
-        if ($usePimcoreSerializer) {
-            $serializer = $this->container->get('pimcore.serializer');
-        } else {
-            $serializer = $this->container->get('serializer');
-        }
-
-        if ($associative) {
-            $context['json_decode_associative'] = true;
-        }
-
-        return $serializer->decode($json, 'json', $context);
-    }
-
-    /**
-     * Translates the given message.
-     *
-     * @param string $id The message id (may also be an object that can be cast to string)
-     * @param array $parameters An array of parameters for the message
-     * @param string|null $domain The domain for the message or null to use the default
-     * @param string|null $locale The locale or null to use the default
-     *
-     * @return string The translated string
-     *
-     * @throws InvalidArgumentException If the locale contains invalid characters
-     */
-    public function trans($id, array $parameters = [], $domain = 'admin', $locale = null)
-    {
-        return $this->translator->trans($id, $parameters, $domain, $locale);
     }
 }
